passed down props and added necessary imports for child components
