+++ src/Components/Attendees.jsx
@@ -1,3 +1,19 @@
-export default function Attendees() {
-  return;
+import Attendee from "./Attendee";
+
+export default function Attendees({ toggleEventAttendees, showAttendees, attendees, event, updateEventAttendance }) {
+  return(
+    <>
+      <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+
+      {showAttendees ? (
+        <div className="attendees">
+          {attendees.map((attendee, index) => (
+            <Attendee attendee={attendee} event={event} updateEventAttendance={updateEventAttendance}/>
+          ))}
+        </div>
+      ) : null}
+    </>
+  );
 }
